requirements.txt
fastapi==0.110.1
openai==1.17.1
pydantic==2.7.0
python-dotenv==1.0.1
uvicorn==0.29.0
zhipuai==2.0.1
